Lump commit of Envision changes to BEMT algorithm
The algorithm is more robust, checking that BEMT solutions are valid before using them.
It now runs when Vx < 0.
The tip/hub induction factors are set to 0 when using the tip/hub correction models (previous versions set them to +/-1 and then occasionally modified them). [jmj wants this fixed!]
The code will not abort if BEMT can't find a valid phi; instead it will turn off BEMT and compute phi geometrically. If this happens at phi=0 or phi=+/-pi/2, no warning will be generated; however if phi is outside epsilon of those values, a warning will be printed on the screen. 
On the first time step, phi will be computed instead of using the uninitialized (phi=0) values. This should give better results in the first few time steps.
I added comments and cleaned up the code.